Allow arbitrary paths and most HTTP methods
This enables specifying arbitrary paths, as well as using all HTTP
request methods supported by Cloud Functions.

Note that when a deployed HTTP Function is invoked, the endpoint name is
NOT available in any of the following attributes on the request object:

- path: Starts with a slash (/) followed by subpath.
- full_path: A question mark (?) plus the URL parameters appended to
  path.
- script_root: Is empty.
- url, base_url: Is the full URL for the request, but does not contain
  the method name.
- url_root: Is the root URL (http://xxx.cloudfunctions.net/)

Effectively, the Cloud Function behaves as though it is hosted at the
root (/).